Swap trip type toggle and activity filter chips positions
Header (TripCreationBar):
- Replace activity filter chips row with Personal/Business/Mixed pills
- Same styling as previous chips (rounded-full, white border on purple)
- tripType stored in state, passed as query param, pre-populated from URL
- Remove INTEREST_CATEGORIES import (no longer needed here)

Form page (new/page.tsx):
- Replace trip type toggle at top with activity filter chips row
- Chips use INTEREST_CATEGORIES, styled as rounded pills
- Selecting chips toggles interest categories in Travel Profile below
- tripType read from URL params (set by search bar) instead of local state
- Remove TRIP_TYPES constant (moved to search bar)
